requirements.txt
streamlit>=1.30.0
google-cloud-documentai>=2.24.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
python-dotenv>=1.0.0
pandas>=2.0.0
